+++ svelte.config.js
@@ -1,4 +1,4 @@
-import adapter from "@sveltejs/adapter-node";
+import adapter from "./scripts/svelte-adapter-node-sqlite.js";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {